+++ src/core/models/__init__.py
@@ -8,7 +8,6 @@ from src.core.models.domain import Domain
 from src.core.models.info_item import InfoItem
 from src.core.models.info_item_rep_spec import InfoItemRepSpec
 from src.core.models.info_item_source import InfoItemSource
-from src.core.models.info_item_source_revision import InfoItemSourceRevision
 from src.core.models.info_source import InfoSource
 from src.core.models.rep_spec import RepSpec
 from src.core.models.source_revision import SourceRevision
@@ -22,7 +21,6 @@ __all__ = [
     "InfoItem",
     "InfoItemRepSpec",
     "InfoItemSource",
-    "InfoItemSourceRevision",
     "InfoSource",
     "RepSpec",
     "SourceRevision",